# Conflicts:
#	speed-dreams/source-2.2.3/src/simulatedDrivingAssistance/backend/DecisionMaker.inl

+++ speed-dreams/source-2.2.3/src/simulatedDrivingAssistance/backend/BlackBoxData.h
@@ -5,9 +5,9 @@
 struct BlackBoxData
 {
 public:
-    BlackBoxData(tCarElt* p_car, tSituation* p_situation, int p_tickCount, tTrackSeg* p_nextSegments, int p_nextSegmentsCount);
+    BlackBoxData(tCarElt* p_car, tSituation* p_situation, unsigned long p_tickCount, tTrackSeg* p_nextSegments, int p_nextSegmentsCount);
     ~BlackBoxData();
     tCarElt Car;
     tSituation Situation;
-    int TickCount;
+    unsigned long TickCount;
 };
