+++ speed-dreams/source-2.2.3/src/simulatedDrivingAssistance/backend/SocketBlackBox.h
@@ -31,7 +31,7 @@ public:
 
 	void DeserializeBlackBoxResults(const char* p_dataReceived, unsigned int p_size, DecisionTuple& p_decisionTuple);
 
-    bool GetDecisions(tCarElt* p_car, tSituation* p_situation, int p_tickCount, DecisionTuple& p_decisions);
+    bool GetDecisions(tCarElt* p_car, tSituation* p_situation, unsigned long p_tickCount, DecisionTuple& p_decisions);
 
     std::vector<std::string> VariablesToReceive;
